+++ src/scvi/nn/_embedding.py
@@ -27,12 +27,7 @@ def _partial_freeze_hook_factory(freeze: int) -> Callable[[torch.Tensor], torch.
 
 
 class Embedding(nn.Embedding):
-    """``EXPERIMENTAL`` Embedding layer with utility methods for extending.
-
-    Notes
-    -----
-    Lifecycle: experimental in v1.2.
-    """
+    """``EXPERIMENTAL`` Embedding layer with utility methods for extending."""
 
     @classmethod
     def extend(