+++ src/scvi/train/__init__.py
@@ -1,7 +1,6 @@
 from ._callbacks import (
     JaxModuleInit,
     LoudEarlyStopping,
-    SaveBestState,
     SaveCheckpoint,
     ScibCallback,
 )
@@ -30,7 +29,6 @@ __all__ = [
     "ClassifierTrainingPlan",
     "TrainRunner",
     "LoudEarlyStopping",
-    "SaveBestState",
     "SaveCheckpoint",
     "ScibCallback",
     "JaxModuleInit",